Выполнит рефакт-г компонента main-page
