+++ src/pages/main-page/components/offers.tsx/utils.ts
