+++ src/pages/main-page/main-page.tsx
@@ -1,16 +1,13 @@
-import { useState } from 'react';
 import cn from 'classnames';
 import Header from '../../components/header/header';
 import Nav from '../../components/nav/nav';
-import Map from '../../components/map/map';
 import { LocationsList } from './components/locations-list/locations-list';
 import { LocationsItem } from './components/locations-item/locations-item';
-import { MainEmpty } from './components/main-empty/main-empty';
-import { MainContent } from './components/main-content.tsx/main-content';
 import { CITIES, TypesPage } from '../../const';
 import { FavoritesListType, TypesPageKeys } from '../../types/types';
 import { useAppSelector } from '../../hooks';
 import { filterOffersByCity } from './utils';
+import { Content } from './components/content/content';
 
 type MainPageProps = {
   isLoggedIn: boolean;
@@ -23,24 +20,14 @@ function MainPage({ isLoggedIn, favorites }: MainPageProps): JSX.Element {
   const cityOffers = filterOffersByCity(offers, currentCity);
   const isEmpty = cityOffers.length === 0;
 
-  const [activeCardId, setActiveCardId] = useState<string | null>(null);
-  const handleCardHover = (id: string | null) => setActiveCardId(id);
 
   const typesPage: TypesPageKeys = TypesPage.Main;
   const mainClasses = cn('page__main page__main--index', {
     ['page__main--index-empty']: isEmpty,
   });
-  const containerClasses = cn('cities__places-container container', {
-    ['cities__places-container--empty']: isEmpty,
-  });
-  const sectionClasses = cn({
-    ['cities__places places']: !isEmpty,
-    ['cities__no-places']: isEmpty,
-  });
 
   return (
     <div className="page page--gray page--main">
-      <span className="visually-hidden">{activeCardId}</span>
       <Header typesPage={typesPage}>
         <Nav
           isLoggedIn={isLoggedIn}
@@ -65,29 +52,11 @@ function MainPage({ isLoggedIn, favorites }: MainPageProps): JSX.Element {
           </section>
         </div>
         <div className="cities">
-          <div className={containerClasses}>
-            <section className={sectionClasses}>
-              {isEmpty ? (
-                <MainEmpty city={currentCity} />
-              ) : (
-                <MainContent
-                  currentCity={currentCity}
-                  onCardHover={handleCardHover}
-                  offers={cityOffers}
-                  typesPage={typesPage}
-                />
-              )}
-            </section>
-            <div className="cities__right-section">
-              {isEmpty || (
-                <Map
-                  offers={cityOffers}
-                  activeCardId={activeCardId}
-                  typesPage={typesPage}
-                />
-              )}
-            </div>
-          </div>
+          <Content
+            cityOffers={cityOffers}
+            currentCity={currentCity}
+            typesPage={typesPage}
+          />
         </div>
       </main>
     </div>
